Windows

+++ pandas/tests/indexes/numeric/test_indexing.py
@@ -402,11 +402,13 @@ class TestGetIndexer:
         idx = Index([1])
 
         result = idx.get_indexer(target)
-        expected = np.array([0, -1])
+        expected = np.array([0, -1], dtype=np.int64)
         tm.assert_numpy_array_equal(result, expected)
 
         result_1, result_2 = idx.get_indexer_non_unique(target)
-        expected_1, expected_2 = np.array([0, -1]), np.array([1])
+        expected_1, expected_2 = np.array([0, -1], dtype=np.int64), np.array(
+            [1], dtype=np.int64
+        )
         tm.assert_numpy_array_equal(result_1, expected_1)
         tm.assert_numpy_array_equal(result_2, expected_2)
